requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
python-dotenv
langgraph
langchain-core
langchain-groq
httpx